WIP - Get attorneys from offices api
Jira ticket: CAMS-283
